Multiple bug fixes and UB fixes
- hk_base: hk_Array_Base default ctor must init members (cabb7d4)
- hk_base: hk_Array<T>::index_of should use correct type (6f7be93)
- hk_base: Correctly set bool running flag in stopwatch (73e6497)
- ivp_controller: Add missed IVP_Constraint_Local members init (201fd15)
- ivp_collision: Zero-initialize collision data in ctor (428c7e2)
- ivp_intern: Fix UB on cast IVP_U_Quat -> IVP_U_Point (746203d)
- ivp_intern: Cache core to speedup set_extra_radius (143e6ba)
- ivp_intern: First setup manager, then assign (e11894b)
- ivp_utility: Fix UB on casts between unrelated types (2e49a17)

+++ havana/havok/hk_base/array/array.inl
@@ -109,7 +109,7 @@ const T& hk_Array<T>::operator() (int i) const
 template <class T>
 hk_array_index hk_Array<T>::index_of( T& t)
 {
-	for (int i = m_n_elems-1; i>=0 ;i--)
+	for (int i = static_cast<int>(m_n_elems)-1; i>=0 ;i--)
 	{
 		if(get_elems()[i]==t)
 		{
